organise the rest of the application files into relevant folders for better scalability
