+++ server/Dockerfile
@@ -9,7 +9,7 @@ COPY . .
 
 # Start both Celery worker and FastAPI app
 CMD /bin/sh -c "\
-  celery -A app.celery_worker.celery_app worker --concurrency=4 --loglevel=INFO & \
+  celery -A app.core.celery_worker.celery_app worker --concurrency=4 --loglevel=INFO & \
   gunicorn --workers=1 --threads=4 --timeout=200 \
     --worker-class=uvicorn.workers.UvicornWorker \
     --bind 0.0.0.0:8000 app.main:app"
