+++ server/app/api/v1/routers/transform.py
@@ -13,7 +13,7 @@ from app.api.v1.models.transform import TransformIn
 from app.api.v1.operations.transform import transform_operation
 from app.api.v1.operations.cleanup import cleanup_operation
 
-from app.config import config as app_config
+from app.core.config import config as app_config
 
 router = APIRouter()
 logger = logging.getLogger("api")
